… go.sum
… staging/src/k8s.io/apiextensions-apiserver/go.sum
… staging/src/k8s.io/apiserver/go.sum
… staging/src/k8s.io/kube-aggregator/go.sum
… staging/src/k8s.io/legacy-cloud-providers/go.sum
… staging/src/k8s.io/sample-apiserver/go.sum
… vendor/modules.txt
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/element.go
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/fromvalue.go
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/path.go
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/set.go
… vendor/sigs.k8s.io/structured-merge-diff/merge/conflict.go
… vendor/sigs.k8s.io/structured-merge-diff/typed/helpers.go
… vendor/sigs.k8s.io/structured-merge-diff/value/value.go