```
$ benchcmp orig.txt new.txt
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2230683       2129184       -4.55%
BenchmarkUpdateNewObject-12     2624417       2379121       -9.35%
BenchmarkRepeatedUpdate-12      1031653       981264        -4.88%
BenchmarkSetToFields-12         24778         23686         -4.41%
BenchmarkFieldsToSet-12         33876         17971         -46.95%
BenchmarkYAMLToTyped-12         68377         64312         -5.94%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      8632           7168           -16.96%
BenchmarkUpdateNewObject-12     11210          8396           -25.10%
BenchmarkRepeatedUpdate-12      3057           2887           -5.56%
BenchmarkSetToFields-12         124            124            +0.00%
BenchmarkFieldsToSet-12         147            93             -36.73%
BenchmarkYAMLToTyped-12         150            152            +1.33%

benchmark                       old bytes     new bytes     delta
BenchmarkApplyNewObject-12      555847        500704        -9.92%
BenchmarkUpdateNewObject-12     659671        560262        -15.07%
BenchmarkRepeatedUpdate-12      255150        221120        -13.34%
BenchmarkSetToFields-12         8974          8369          -6.74%
BenchmarkFieldsToSet-12         12897         4688          -63.65%
BenchmarkYAMLToTyped-12         27625         27384         -0.87%
```

```release-note
NONE
```